+++ plone/app/contentlisting/contentlisting.py
@@ -26,6 +26,9 @@ class ContentListing(object):
     def __getitem__(self, index):
         """`x.__getitem__(index)` <==> `x[index]`
         """
+        if isinstance(index, slice):
+            return IContentListing(
+                self._basesequence[index.start:index.stop:index.step])
         return IContentListingObject(self._basesequence[index])
 
     def __len__(self):
@@ -66,6 +69,10 @@ class ContentListing(object):
         """`x.__eq__(other)` <==> `x == other`"""
         raise NotImplementedError
 
+    def __hash__(self):
+        """`x.__hash__()`"""
+        raise NotImplementedError
+
     def __ne__(self, other):
         """`x.__ne__(other)` <==> `x != other`"""
         raise NotImplementedError
@@ -93,7 +100,8 @@ class ContentListing(object):
     def __getslice__(self, i, j):
         """`x.__getslice__(i, j)` <==> `x[i:j]`
         Use of negative indices is not supported.
-        Deprecated since Python 2.0 but still a part of `UserList`.
+        No longer used in Python 3, but still part of
+        zope.interface.interfaces.IReadSequence
         """
         return IContentListing(self._basesequence[i:j])
 
@@ -108,6 +116,9 @@ class BaseContentListingObject(object):
         """For comparing two contentlistingobject"""
         other = IContentListingObject(other)
         return self.uuid() == other.uuid()
+
+    def __hash__(self):
+        return hash(self.uuid())
 
     def ContentTypeClass(self):
         # A normalised type name that identifies the object in listings.